+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/types/ObjectType.java
@@ -20,6 +20,27 @@ public class ObjectType extends BuiltinType {
         this.typeVariableScope.declare(List.of(TokenType.CONST), "init", new Init());
         this.typeVariableScope.declare(List.of(TokenType.CONST), "toString", new ToString());
         this.typeVariableScope.declare(List.of(TokenType.CONST), "equals", new Equals());
+        this.typeVariableScope.declare(List.of(TokenType.CONST), "fromNBT", new FromNBT());
+    }
+
+    private static class Equals extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            this.validate("equals", arguments, List.of(ObjectType.TYPE));
+
+            BuiltinClass object = arguments.get(0);
+
+            return new BooleanInstance(this.boundClass.equals(object));
+        }
+    }
+
+    private static class FromNBT extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            this.validate("fromNBT", arguments, List.of(ObjectType.TYPE));
+
+            return this.boundClass.fromNBT(BuiltinClass.fromNbtElement(arguments.get(0).toNBT()));
+        }
     }
 
     private static class Init extends BuiltinMethod {
@@ -37,17 +58,6 @@ public class ObjectType extends BuiltinType {
             this.validate("toString", arguments);
 
             return new StringInstance(this.boundClass.toString());
-        }
-    }
-
-    private static class Equals extends BuiltinMethod {
-        @Override
-        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
-            this.validate("equals", arguments, List.of(ObjectType.TYPE));
-
-            BuiltinClass object = arguments.get(0);
-
-            return new BooleanInstance(this.boundClass.equals(object));
         }
     }
 }